pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "ros_nerf"
version = "0.0.9"

dependencies = [
    # the pypi version of ros_numpy has a broken version of point_cloud2.py so we want this specific commit
    "ros_numpy@git+https://github.com/eric-wieser/ros_numpy@74879737c8648f48adb507a5bdf4e51c0d194124",
    "opencv-python",
    "catkin_pkg",
    "empy==3.3.4",
    "numpy",
    "nerfstudio==1.1.3",
    "catkin_tools",
    "rospkg",
    "pyyaml",
    "pillow"
]

[tool.setuptools]
packages = ["ros_nerf"]

[project.entry-points.'nerfstudio.dataparser_configs']
ros-data = "ros_nerf.data.ros_dataparser:ROSDataparser"

[project.scripts]
ns-ros = "ros_nerf.scripts.ros_runner:entrypoint"
ns-ros-save = "ros_nerf.scripts.ros_saver:entrypoint"